backend/youtube/requirements.txt
boto3
psycopg2-binary
pytube @ git+https://github.com/capynotes/pytube